+++ R/add_gml_mids.R
@@ -5,9 +5,8 @@
 #'
 #' @return \code{add_gml_mids()} takes a dyad-year or state-year data frame and adds dyad-year dispute information
 #' from the GML MID data. If the data are dyad-year, the return is a laundry list of information about onsets, ongoing conflicts,
-#' and assorted participant- and dispute-level summaries. If the data are state-year, the function returns information about ongoing
-#' disputes (and onsets) and whether there were any ongoing disputes (and onsets) the state initiated. A character string of dispute
-#' numbers active in the observation year is returned too.
+#' and assorted participant- and dispute-level summaries. If the data are state-year or leader-year, the function returns information about ongoing
+#' disputes (and onsets) and whether there were any ongoing disputes (and onsets) the state (or leader) initiated.
 #'
 #' @details Dyads are capable of having multiple disputes in a given year, which can create a problem
 #' for merging into a complete dyad-year data frame. Consider the case of France and Italy in 1860, which
@@ -20,13 +19,22 @@
 #' was bilateral and the dispute did not escalate beyond an attack, the state on Side A initiated the dispute. For multilateral MIDs, these
 #' conditions still hold at least for originators. However, there is *considerable* difficulty for cases where 1) participant-level summaries
 #' suggested actions at the level of clash or higher, 2) the participant was a joiner and not an originator. The effort required to
-#' flesh this out is enormous, and perhaps forthcoming in a future update. For now, the simple determination is all states on Side A that
-#' were originators are all initiators, even though there are conceivable cases where, for example, a participant on Side B initiated
-#' a course of action as a joiner (e.g. the Sudetenland crisis comes to mind here). Everyone else is not an originator.
+#' flesh this out is enormous, and perhaps forthcoming in a future update.
+#'
+#' \code{add_gml_mids()} allows you to make one of three judgment
+#' calls here (see the arguments section of the documentation). If it were my call to make, I would say you should probably use the option
+#' \code{"sidea-all-joiners"}. My review of the MID data with Doug Gibler suggests most states that join a dispute are not roped into
+#' a conflict (i.e. targeted by some other state) after the first incident. They routinely initiate their entry into the conflict, which is
+#' what this concept of "initiation" is supposed to capture in the literature. There are no doubt cases where some third state is brought into
+#' the dispute by the actions of some other state even as the original MID coding rules place a high barrier on coding that type of dispute entry.
+#' However, the time required to individually assess whether a state initiated their entry into a MID under something other than the
+#' simplest of cases (e.g. bilateral cases where the highest participant action fell short of a clash) would be too time-consuming. It would
+#' require an audit of almost half of all participant-level summaries in the data. In a forthcoming publication, Gibler and Miller offer
+#' excellent coverage here with a new data set on militarized events. However, this would include only confrontations after World War II.
 #'
 #' @author Steven V. Miller
 #'
-#' @param data a dyad-year data frame (either "directed" or "non-directed")
+#' @param data a dyad-year data frame (either "directed" or "non-directed"), a state-year data frame, or a leader-year data frame
 #' @param keep an optional parameter, specified as a character vector, applicable to just the dyad-year data, and
 #' passed to the function in a \code{select(one_of(.))} wrapper. This
 #' allows the user to discard unwanted columns from the directed dispute data so that the output does not consume
@@ -36,6 +44,14 @@
 #' The user is free to keep or discard anything else they see fit.
 #'
 #' If \code{keep} is not specified in the function, the ensuing output returns everything.
+#'
+#' @param init how should initiators be coded? Applicable only to state-year and leader-year data. This parameter accepts one of
+#' three possible values (\code{"sidea-orig"}, \code{"sidea-with-joiners"}, \code{"sidea-all-joiners"}). \code{"sidea-orig"} = a state initiates a MID (which
+#' appears as a summary return in the output) if the state was on Side A at the onset of the dispute. \code{"sidea-with-joiners"} = a state
+#' initiates a MID (which appears as a summary return in the output) if the state was on Side A at the onset of the dispute or if the
+#' state joined the MID on Side A. \code{"sidea-all-joiners"} = a state initiates a MID (which appears as a summary return in the output) if the
+#' state was on Side A at the onset of the dispute or if it joined at any point thereafter. See details section for more discussion. The
+#' default is  \code{"sidea-all-joiners"}.
 #'
 #' @references
 #'
@@ -55,7 +71,11 @@
 #'
 
 
-add_gml_mids <- function(data, keep) {
+add_gml_mids <- function(data, keep, init = "sidea-all-joiners") {
+
+  if (init %nin% c("sidea-all-joiners", "sidea-with-joiners", "sidea-orig")) {
+    stop("init must be one of 'sidea-all-joiners', 'sidea-with-joiners', or 'sidea-orig'. You may have mispelled something. This argument applies only to state-year or leader-year analyses.")
+  }
 
   if (length(attributes(data)$ps_data_type) > 0 && attributes(data)$ps_data_type == "dyad_year") {
 
@@ -113,7 +133,56 @@ add_gml_mids <- function(data, keep) {
 
   } else  if (length(attributes(data)$ps_data_type) > 0 && attributes(data)$ps_data_type == "leader_year") {
 
-    print("It's coming. Trust me.")
+    leaderdays <- create_leaderdays()
+
+    gml_part %>%
+      filter(allmiss_leader_start == 0 & allmiss_leader_end == 0) %>%
+      mutate(stdate = as.Date(paste0(styear,"/",stmon,"/", dummy_stday)),
+             enddate = as.Date(paste0(endyear,"/",endmon,"/", dummy_endday))) %>%
+      rowwise() %>%
+      mutate(date = list(seq(stdate, enddate, by = "1 day")),
+             gmlmidonset = list(ifelse(date == min(date), 1, 0))) %>%
+      unnest(c(.data$date, .data$gmlmidonset)) %>%
+      mutate(gmlmidongoing = 1) -> hold_this
+
+    if (init == "sidea-orig") {
+      hold_this %>%
+        mutate(gmlmidongoing_init = ifelse(gmlmidongoing == 1 & (sidea == 1 & orig == 1), 1, 0),
+               gmlmidonset_init = ifelse(gmlmidonset == 1 & gmlmidongoing_init == 1, 1, 0)) %>%
+        select(dispnum:ccode, obsid_start,  sidea, orig, date:gmlmidonset_init) -> hold_this
+
+    } else if (init == "sidea-with-joiners") {
+
+      hold_this %>%
+        mutate(gmlmidongoing_init = ifelse(gmlmidongoing == 1 & (sidea == 1 | ( orig == 0 & sidea == 1 )), 1, 0),
+               gmlmidonset_init = ifelse(gmlmidonset == 1 & gmlmidongoing_init == 1, 1, 0)) %>%
+        select(dispnum:ccode, obsid_start, sidea, orig, date:gmlmidonset_init) -> hold_this
+
+    } else if (init == "sidea-all-joiners") {
+
+      hold_this %>%
+        mutate(gmlmidongoing_init = ifelse(gmlmidongoing == 1 & (sidea == 1 | ( orig == 0)), 1, 0),
+               gmlmidonset_init = ifelse(gmlmidonset == 1 & gmlmidongoing_init == 1, 1, 0)) %>%
+        select(dispnum:ccode, obsid_start,  sidea, orig, date:gmlmidonset_init) -> hold_this
+
+    }
+
+    hold_this %>%
+      left_join(leaderdays, .) %>%
+      mutate(year = .pshf_year(date)) %>%
+      group_by(obsid, year) %>%
+      summarize(gmlmidongoing = sum(gmlmidongoing, na.rm=T),
+                gmlmidonset = sum(gmlmidonset, na.rm=T),
+                gmlmidongoing_init = sum(gmlmidongoing_init, na.rm=T),
+                gmlmidonset_init = sum(gmlmidonset_init, na.rm=T)) -> hold_this
+
+
+    hold_this %>%
+      mutate_at(vars("gmlmidongoing", "gmlmidonset", "gmlmidongoing_init", "gmlmidonset_init"), ~ifelse(. >= 1, 1, 0)) -> hold_this
+
+    data %>%
+      left_join(., hold_this) -> data
+
 
   } else {
     stop("add_gml_mids() requires a data/tibble with attributes$ps_data_type of state_year or dyad_year. Try running create_dyadyears() or create_stateyears() at the start of the pipe.")